% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{check_rnaseqcomp}
\alias{check_rnaseqcomp}
\title{Sanity Check of S4 rnaseqcomp class}
\usage{
check_rnaseqcomp(object)
}
\arguments{
\item{object}{A object of S4 rnaseqcomp class}
}
\value{
TRUE, or character if error happens.
}

