% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrixFilter.R
\name{matrixFilter}
\alias{matrixFilter}
\title{Quantification Table Filtering}
\usage{
matrixFilter(quantData, repInfo, evaluationFeature, calibrationFeature,
  unitReference, unitCutoff = 1)
}
\arguments{
\item{quantData}{A matrix of quantifications with rows by
features (genes, transcripts, junctions or exons) and columns
by pipelines each with 2 replicates. Missing data is allowed
as NA.}

\item{repInfo}{A factor documenting quantification pipeline
names corresponding to the columns of \code{quantData}.}

\item{evaluationFeature}{A logical vector corresponding to the
rows of \code{quantData}, providing which features should be
considered for downstream evaluation, e.g. protein coding genes.}

\item{calibrationFeature}{A logical vector corresponding to the
rows of \code{quantData}, providing which features should be
considered as calibration reference, e.g. house keeping genes.}

\item{unitReference}{A logical vector corresponding to the columns
of \code{quantData}, providing to which columns all units should be
unified and given threshold \code{unitCutoff} should be applied.
If multiple columns are chosen, it is important that these columns
have the same units (e.g. FPKM).}

\item{unitCutoff}{A numeric threshold for signal cutoff at given
quantification table columns by \code{unitReference} (default: 1).
Same effects cutoff are estimated for other columns accordingly.}
}
\value{
A \code{rnaseqcomp} S4 class object
\item{quantData}{A filtered table for downstream analysis.}
\item{repInfo}{A factor documenting quantification pipeline
names corresponding to the columns of \code{quantData}.}
\item{refMed}{A numeric vector giving the log scale medians
of calibration reference.}
\item{scaler}{A number providing the scales for downstream
detrended signal.}
}
\description{
This is the function to do any pre-filtering or
pre-processing analysis for downstream benchmark estimation
and visualization. Pre-filtering includes row selection (e.g.
protein coding genes) of quantification table ; pre-processing
includes calculation on a set of rows as calibration reference
(e.g. house keeping genes) across different quantification
pipelines, estimation of confidence thresholds on all the
pipelines based on given thresholds from selected pipelines.
}
\details{
In the functions \code{plotMAD} and \code{plotNE}, detrended
signals with value 0 will be at the same level as value 1 for
giving pipelines by \code{unitReference}.
}
\examples{
data(encodeCells)
evaluationFeature <- encodeCells$genemeta$type == "protein_coding"
calibrationFeature <- encodeCells$genemeta$housekeeping
unitReference <- grepl("Cufflinks",encodeCells$repInfo)
dat <- matrixFilter(encodeCells$gm12878,encodeCells$repInfo,
evaluationFeature,calibrationFeature,unitReference)
}

