% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotCAT.R
\name{plotCAT}
\alias{plotCAT}
\title{CAT Plots of Differential Express Features}
\usage{
plotCAT(dat1, dat2, constant = NULL, otherFC = NULL, infinity = FALSE,
  step = 5L, type = "l", lwd = 2, col = NULL, lty = 1, xlim = c(20L,
  500L), ylim = c(0, 1), xlab = "Size of List",
  ylab = "Proportion in Common", cex.leg = 0.6, ...)
}
\arguments{
\item{dat1,dat2}{\code{rnaseqcomp} S4 class objects for two
conditions. \code{dat1} and \code{dat2} should have the same
size of slot \code{quantData} and the same \code{repInfo}.}

\item{constant}{A numeric constant that can be added to
quantifications before fold changes calculation
(default: NULL).}

\item{otherFC}{A numeric vector of fold changes by
other independent technology such as microarray, with each elements
corresponding to rows of \code{quantData} slot in \code{dat1}
or \code{dat2}. Missing data NA allowed. (default: NULL)}

\item{infinity}{A logical indicator that specify if fold change
of infinity should be considered. Functional only if \code{constant}
is not a positive number. (default: FALSE)}

\item{step}{Plot steps on x-axis (default: 5).}

\item{type}{Plot types (default: 'l').}

\item{lwd}{Plot line weights (default: 2).}

\item{col}{Plot colors (default: NULL, colors are assigned
by package \code{RColorBrewer}).}

\item{lty}{Plot line styles (default: 1).}

\item{xlim}{Plot limits of x-axis (default: c(20, 500)).}

\item{ylim}{Plot limits of y-axis (default: c(0, 1)).}

\item{xlab}{Plot label of x-axis (default: 'Size of List').}

\item{ylab}{Plot label of y-axis (default: 'Proportion in Common').}

\item{cex.leg}{Legend size (default: 0.6).}

\item{...}{Other parameters for base function \code{plot}.}
}
\value{
\item{CAT plot}{CAT plots for all the quantification pipelines.}
\item{Precision or Accuracy}{A numeric vector of pipeline
precision or accuracy, depending on whether \code{otherFC} is
provided.}
}
\description{
For each pipeline, 2 biological conditions
(cell lines) each with 2 quantification replicates are
compared and fold changes of each replicate are calculated.
Then, CAT plots between replicates (precision) or between
mean of replicates and other technology (accuracy).
}
\examples{
data(encodeCells)
evaluationFeature <- encodeCells$genemeta$type == "protein_coding"
calibrationFeature <- encodeCells$genemeta$housekeeping
unitReference <- grepl("Cufflinks",encodeCells$repInfo)
dat1 <- matrixFilter(encodeCells$gm12878,encodeCells$repInfo,
    evaluationFeature,calibrationFeature,unitReference)
dat2 <- matrixFilter(encodeCells$k562,encodeCells$repInfo,
    evaluationFeature,calibrationFeature,unitReference)

plotCAT(dat1,dat2)
plotCAT(dat1,dat2,infinity=TRUE)

genes <- encodeCells$genemeta[encodeCells$genemeta$type ==
    "protein_coding", 1]
otherFC <- encodeCells$arrayFC[match(genes,names(encodeCells$arrayFC))]
plotCAT(dat2,dat1,otherFC=otherFC)
plotCAT(dat2,dat1,constant=1,otherFC=otherFC)
}

