% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotMAD.R
\name{plotMAD}
\alias{plotMAD}
\title{Plots of Median Absolute Deviation}
\usage{
plotMAD(dat, type = "l", lwd = 2, col = NULL, lty = 1, xlim = NULL,
  ylim = NULL, cex.leg = 0.6, xlab = "Detrended logSignal",
  ylab = "MAD", ...)
}
\arguments{
\item{dat}{A \code{rnaseqcomp} S4 class object.}

\item{type}{Plot types (default: 'l').}

\item{lwd}{Plot line weights (default: 2).}

\item{col}{Plot colors (default: NULL, colors are assigned
by package \code{RColorBrewer}).}

\item{lty}{Plot line styles (default: 1).}

\item{xlim}{Plot limits of x-axis (default: NULL, limits are
estimated automatically).}

\item{ylim}{Plot limits of y-axis (default: NULL, limits are
estimated automatically).}

\item{cex.leg}{Legend size (default: 0.6).}

\item{xlab}{Plot label of x-axis
(default: 'Detrended logSignal').}

\item{ylab}{Plot label of y-axis (default: 'MAD').}

\item{...}{Other parameters for base function \code{plot}.}
}
\value{
\item{MAD plot}{MAD plots for all the quantification pipelines.}
\item{MAD}{A numeric vector of median absolute deviations.}
}
\description{
For each pipeline, two quantification replicates
are compared and log scale absolute deviations of signals are
calculated. Then, loess smooth on absolute deviation are
plotted stratified by detrended log signals.
}
\examples{
data(encodeCells)
evaluationFeature <- encodeCells$genemeta$type == "protein_coding"
calibrationFeature <- encodeCells$genemeta$housekeeping
unitReference <- grepl("Cufflinks",encodeCells$repInfo)
dat <- matrixFilter(encodeCells$gm12878,encodeCells$repInfo,
evaluationFeature,calibrationFeature,unitReference)
plotMAD(dat)
}

