% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{check_rnaseqcomp}
\alias{check_rnaseqcomp}
\title{Sanity Check of S4 rnaseqcomp class}
\usage{
check_rnaseqcomp(object)
}
\arguments{
\item{object}{A object of S4 rnaseqcomp class}
}
\value{
TRUE, or charater if error happens.
}
\description{
Sanity Check of S4 rnaseqcomp class
}

