% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrixFilter.R
\name{matrixFilter}
\alias{matrixFilter}
\title{Quantification Table Filtering}
\usage{
matrixFilter(quantData, repInfo, txFIdx, hkIdx, unitFIdx, unitCut = 1)
}
\arguments{
\item{quantData}{A matrix of quantifications with rows by genes,
transcriptsm, exons etc. and columns by pipelines each with 2
replicates. Missing data is allowed as NA.}

\item{repInfo}{A factor documenting quantification pipeline
names correponding to the columns of \code{quantData}.}

\item{txFIdx}{A logical vector corresponding to the rows of
\code{quantData}, providing which rows should be considered
for downsteam analyis, e.g. protein coding genes.}

\item{hkIdx}{A logical vector corresponding to the rows of
\code{quantData}, providing which rows should be considered as
calibration reference, e.g. house keeping genes.}

\item{unitFIdx}{A logical vector corresponding to the columns
of \code{quantData}, providing which columns (i.e. pipelines)
should be applied for given threshold \code{unitCut}.}

\item{unitCut}{A numeric threshold for signal cutoff at given
quantification table columns by \code{unitFidx} (default: 1).
The effects cutoff are estimated for other columns accordingly.}
}
\value{
A \code{rnaseqcomp} S4 class object
\item{quantData}{A filtered table for downsteam analysis.}
\item{repInfo}{A factor documenting quantification pipeline
names correponding to the columns of \code{quantData}.}
\item{hkmed}{A numeric vector giving the log scale medians
of calibration reference.}
\item{scaler}{A number providing the scales for downstream
detrended signal.}
}
\description{
This is the function to do any pre-filtering or
pre-processing analysis for dowstreaming benchmark estimation
and visualization. Pre-filtering includes row selection (e.g.
protein coding genes) of quantification table ; pre-processing
inlcudes calculation on a set of rows as calibration reference
(e.g. house keeping genes) across different quantification
pipelines, estimation of confidence thresholds on all the
pipelins based on given thresholds from selected pipelines.
}
\details{
In the functions \code{plotMAD} and \code{plotNE}, detrended
signals with value 0 will be at the same level as value 1 for
giving pipelines by \code{unitFIdx}.
}
\examples{
data(encodeCells)
txFIdx <- genemeta$type == "protein_coding"
hkIdx <- genemeta$housekeeping
unitFIdx <- grepl("Cufflinks",repInfo)
dat <- matrixFilter(gm12878,repInfo,txFIdx,hkIdx,unitFIdx)
}

