% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotNE.R
\name{plotNE}
\alias{plotNE}
\title{Plots of Express And Non-express Features}
\usage{
plotNE(dat, step = 0.1, type = "l", lwd = 2, col = NULL, lty = 1,
  xlim = NULL, ylim = NULL, cex.leg = 0.6, xlab = "Detrended logSignal",
  ylab = "Proportion of discordant calls", ...)
}
\arguments{
\item{dat}{A \code{rnaseqcomp} S4 class object.}

\item{step}{Plot steps on x-axis.}

\item{type}{Plot types (default: 'l').}

\item{lwd}{Plot line weights (default: 2).}

\item{col}{Plot colors (default: NULL, colors are assigned
by package \code{RColorBrewer}).}

\item{lty}{Plot line styles (default: 1).}

\item{xlim}{Plot limits of x-axis (default: NULL, limits are
estimated automatically).}

\item{ylim}{Plot limits of y-axis (default: NULL, limits are
estimated automatically).}

\item{cex.leg}{Legend size (default: 0.6).}

\item{xlab}{Plot label of x-axis
(default: 'Detrended logSignal').}

\item{ylab}{Plot label of y-axis
(default: 'Proportion of discordant calls').}

\item{...}{Other parameters for base function \code{plot}.}
}
\value{
\item{Either-or-express plot}{Plots for all the quantification
pipelines.}
\item{matrix}{A proportion matrix of express, non-express and
either-or-express for all pipelines.}
}
\description{
For each pipeline, two quantification replicates
are compared and proportions of both-express, both-non-express
and either-or-express features are calculated. Then, reverse
proportion accumulation for either-or-express features are
plotted stratified by detrended log signals.
}
\examples{
data(encodeCells)
evaluationFeature <- encodeCells$genemeta$type == "protein_coding"
calibrationFeature <- encodeCells$genemeta$housekeeping
unitReference <- grepl("Cufflinks",encodeCells$repInfo)
dat <- matrixFilter(encodeCells$gm12878,encodeCells$repInfo,
    evaluationFeature,calibrationFeature,unitReference)
plotNE(dat)
}

