% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNE.R
\name{plotNE}
\alias{plotNE}
\title{Estimate And Plot Express And Non-express Features}
\usage{
plotNE(dat, steps = seq(-0.5, 12, 0.5), Ks = 0:3, pchK = seq_along(Ks) -
  1, plotcell = 1, lwd = 2, col = NULL, lty = 1, cex.leg = 1,
  xlim = NULL, ylim = NULL, xlab = "\% of detrended logSignal below K",
  ylab = "\% of disagreement between replicates", ...)
}
\arguments{
\item{dat}{A \code{rnaseqcomp} S4 class object.}

\item{steps}{A numeric vector specifying log-scale cutoffs to be used
for calculation and plotting. (default: seq(-0.5, 12, 0.5))}

\item{Ks}{A numeric vector specifying which cutoffs to be highlighted
and to which the reported proportions to be corresponding.}

\item{pchK}{Plot styles of highlight points corresponding
to \code{Ks}. (default: seq_along(Ks) - 1)}

\item{plotcell}{Either 1 or 2 indicating which cell line
will be plotted. This won't affect estimation for both
cell lines. (default: 1)}

\item{lwd}{Plot line weights (default: 2).}

\item{col}{Plot colors (default: NULL, colors are assigned
by package \code{RColorBrewer}).}

\item{lty}{Plot line styles (default: 1).}

\item{cex.leg}{Legend size (default: 1).}

\item{xlim}{Plot limits of x-axis (default: NULL, limits are
estimated automatically).}

\item{ylim}{Plot limits of y-axis (default: NULL, limits are
estimated automatically).}

\item{xlab}{Plot label of x-axis.
(default: 'Proportion of detrended logSignal below K')}

\item{ylab}{Plot label of y-axis.
(default: 'Proportion of disagreement between replicates')}

\item{...}{Other parameters for base function \code{plot}.}
}
\value{
\item{plot}{NE plots of quantification pipelines for
selected cell line by \code{plotcell}.}
\item{NE}{A list of sub-list with each sub-list corresponding to
one cell line. Each element of sub-list is one 2*n matrix for
one pipeline, where \code{n=length(Ks)}. The first row of
matrix gives the proportion
of disagreement and the second row gives the proportion of both
replicates under (non-express) correspoinding cutoff \code{Ks}.}
}
\description{
For each cell line, any compared two replicates
might have a portion of transcripts that express in one replicate
but not the other, depending on what cutoff is used to define
non-express. This function estimate and plot the proportion of
disagreement using multiple cutoffs. Average is used when multiple
two-replicate comparisons included.
}
\examples{
data(simdata)
condInfo <- factor(simdata$samp$condition)
repInfo <- factor(simdata$samp$replicate)
evaluationFeature <- rep(TRUE, nrow(simdata$meta))
calibrationFeature <- simdata$meta$house & simdata$meta$chr == 'chr1'
unitReference <- 1
dat <- signalCalibrate(simdata$quant, condInfo, repInfo, evaluationFeature,
calibrationFeature, unitReference, calibrationFeature2 = calibrationFeature)
plotNE(dat)
}

