% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSD.R
\name{plotSD}
\alias{plotSD}
\title{Estimate And Plot Median Standard Deviation}
\usage{
plotSD(dat, constant = 0.5, loessspan = 1/3, thresholds = c(1, 6),
  plotcell = 1, lwd = 2, col = NULL, lty = 1, cex.leg = 1,
  xlim = NULL, ylim = NULL, xlab = "Detrended logSignal", ylab = "SD",
  ...)
}
\arguments{
\item{dat}{A \code{rnaseqcomp} S4 class object.}

\item{constant}{A numeric pseudo-constant to be added on all
the signals before transferred to log scale. (default: 0.5)}

\item{loessspan}{A numeric number indicating span used
for loess smooth. Details see
\code{loess.smooth} function. (Default: 1/3)}

\item{thresholds}{A vector of two numbers define cutoffs for
three levels of detreded log signals. (default: c(1, 6))}

\item{plotcell}{Either 1 or 2 indicating which cell line
will be plotted. This won't affect estimation for both
cell lines. (default: 1)}

\item{lwd}{Plot line weights (default: 2).}

\item{col}{Plot colors (default: NULL, colors are assigned
by package \code{RColorBrewer}).}

\item{lty}{Plot line styles (default: 1).}

\item{cex.leg}{Legend size (default: 1).}

\item{xlim}{Plot limits of x-axis (default: NULL, limits are
estimated automatically).}

\item{ylim}{Plot limits of y-axis (default: NULL, limits are
estimated automatically).}

\item{xlab}{Plot label of x-axis
(default: 'Detrended logSignal').}

\item{ylab}{Plot label of y-axis (default: 'SD').}

\item{...}{Other parameters for base function \code{plot}.}
}
\value{
\item{plot}{SD plots of quantification pipelines for
selected cell line by \code{plotcell}.}
\item{SD}{A list of two matrices of median standard deviations.}
}
\description{
For each cell line in each pipeline,
the standard deviation of detrend logsignals are calculated
for individual features. Then, loess smooth on standard
deviation are plotted stratified by detrended log signals
for select cell line. The median of standard deviation at
three different levels of detrend logsignals are reported.
}
\examples{
data(simdata)
condInfo <- factor(simdata$samp$condition)
repInfo <- factor(simdata$samp$replicate)
evaluationFeature <- rep(T, nrow(simdata$meta))
calibrationFeature <- simdata$meta$house & simdata$meta$chr == 'chr1'
unitReference <- 1
dat <- signalCalibrate(simdata$quant, condInfo, repInfo, evaluationFeature,
calibrationFeature, unitReference, calibrationFeature2 = calibrationFeature)
plotSD(dat)
}

