% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{rnaseqcomp-class}
\alias{rnaseqcomp-class}
\title{rnaseqcomp}
\description{
This is a S4 class to organize data ready for benchmark summarization.
There are 5 S3 objects inside this class. \code{quantData} documents a
list of data matrices ready for evaluation by functions \code{plotSD},
\code{plotNE}, \code{plot2TX} or \code{plotROC}. \code{condInfo} is a
factor corresponding to columns of \code{quantData} matrices, indicating
to which cell lines each sample belongs. \code{repInfo} is a factor
corresponding to columns of \code{quantData} matrices indicating replicate
information. \code{repInfo} is a legacy from previous versions, and
doesn't have too much meanings in current version. \code{refMed} is the
median log2 signal of calibration references. \code{scaler} is a number
that point to the median log2 signal of reference pipeline. \code{refMed}
and \code{scaler} were used to calibrate and generate \code{quantData}.
}
